[*] [examples.me_Counters] Okay, looks like it works now..
There was logic bug in prev version:

  TBool IsWrapped = AddTimestamp(&EndTs, DeltaTs);

instead of

  TBool IsWrapped = !AddTimestamp(&EndTs, DeltaTs);

But version before code was just stucked in wait loop (time was stopped?).

"volatile" seems to solve this problem (although it may work without
"volatile" but GCC can mess thing up with more probability).

"volatile" for struct is fucking funny: you can't read and assign from/to
volatile record but can read/assign it's integer fields. Concepts are
not for Germans.